requirements.txt
fastapi
uvicorn
torch
torchvision
opencv-python
numpy
pillow
ultralytics
python-multipart
python-dotenv
google-genai
